feat(counter): convert counter to functional component & add example of class and constructor
